% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classify2.R
\name{classify2}
\alias{classify2}
\title{runs the classify() function}
\usage{
classify2(file = "classify_config.txt", ...)
}
\arguments{
\item{file}{path to the saved configuration}

\item{...}{any other parameters to be passed to the
\code{\link{classify}} function}
}
\description{
Allows to read configuration from file without using GUI.
}

