% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compute.freqs.R
\name{compute.freqs}
\alias{compute.freqs}
\title{converts stylo.corpus into features frequency matrix}
\usage{
\method{compute}{freqs}(corpus, absent.sensitive = TRUE, relative = TRUE)
}
\arguments{
\item{corpus}{stylo.corpus object (obtained from
\code{\link{parse.corpus}})}

\item{absent.sensitive}{see \code{\link{make.table.of.frequencies}}}

\item{relative}{should frequencies be relative to text length or absolute}
}
\description{
converts stylo.corpus into features frequency matrix
}
\seealso{
\code{\link{make.table.of.frequencies}},
  \code{\link{parse.corpus}}
}

