% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/db.connect.R
\name{db.connect}
\alias{db.connect}
\title{establishes connection with the corpus database}
\usage{
db.connect(dbname = "styloworkshop", host = "helios.arz.oeaw.ac.at",
  user = "styloworkshop", password = "95CrQCvwP6", ...)
}
\arguments{
\item{dbname}{database name}

\item{host}{database host}

\item{user}{user name}

\item{password}{user password}

\item{...}{other parameters to be passed to dplyr::src_postgres}
}
\value{
connection
}
\description{
Connection is set up in the \code{.src} variable of the user environment, so
it can be used by \code{\link{get.texts}} automatically.
}

