% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.texts.R
\name{get.texts}
\alias{get.texts}
\title{fetches all texts from corpus database}
\usage{
get.texts()
}
\description{
Fetched text can then be filtered and writed to disk in a format supported by the stylo package.
}
\seealso{
\code{\link{write.corpus}}
}

