% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.values.R
\name{get.values}
\alias{get.values}
\title{returns values frequency list for a given column or columns set}
\usage{
get.values(data, columns)
}
\arguments{
\item{data}{set of texts (obtained from \code{\link{get.texts}})}

\item{columns}{column name (or vector of column names)}
}
\value{
data.frame with frequency list
}
\description{
returns values frequency list for a given column or columns set
}
\examples{
\dontrun{
  # connect to the texts database
  db.connect()

  # fetch and filter texts
  allTexts = get.texts()

  # get frequency list for all authors
  get.values(allTexts, 'author')
}
}
\seealso{
\code{\link{get.texts}}
}

