% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oppose2.R
\name{oppose2}
\alias{oppose2}
\title{runs the oppose() function}
\usage{
oppose2(file = "oppose_config.txt", ...)
}
\arguments{
\item{file}{path to the saved configuration}

\item{...}{any other parameters to be passed to the
\code{\link{oppose}} function}
}
\description{
Allows to read configuration from file without using GUI.
}

