% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.config.R
\name{read.config}
\alias{read.config}
\title{prepares stylo functions parameters}
\usage{
read.config(file, ...)
}
\arguments{
\item{file}{path to the file with a stored condig}

\item{...}{manualy set parameters}
}
\value{
list of parameters
}
\description{
Used internally by \code{\link{stylo2}}, \code{\link{classify2}},
\code{\link{rolling.classify2}}, \code{\link{rolling.delta2}},
\code{\link{oppose2}}.
}

