% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rolling.classify2.R
\name{rolling.classify2}
\alias{rolling.classify2}
\title{runs the rolling.classify() function}
\usage{
rolling.classify2(file = "rolling.classify_config.txt", ...)
}
\arguments{
\item{file}{path to the saved configuration}

\item{...}{any other parameters to be passed to the
\code{\link{rolling.classify}} function}
}
\description{
Allows to read configuration from file without using GUI.
}

