% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rolling.delta2.R
\name{rolling.delta2}
\alias{rolling.delta2}
\title{runs the rolling.delta() function}
\usage{
rolling.delta2(file = "rolling.delta_config.txt", ...)
}
\arguments{
\item{file}{path to the saved configuration}

\item{...}{any other parameters to be passed to the
\code{\link{rolling.delta}} function}
}
\description{
Allows to read configuration from file without using GUI.
}

