% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stylo2.R
\name{stylo2}
\alias{stylo2}
\title{runs the stylo() function}
\usage{
stylo2(file = "stylo_config.txt", ...)
}
\arguments{
\item{file}{path to the saved configuration}

\item{...}{any other parameters to be passed to the
\code{\link{stylo}} function}
}
\description{
Allows to read configuration from file without using GUI.
}

