% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write.corpus.R
\name{write.corpus}
\alias{write.corpus}
\title{writes texts to a disk in a format supported by the stylo package}
\usage{
write.corpus(data, groupBy1 = character(), groupBy2 = character(),
  directory = "corpus", sample = 1, limit = 10^7)
}
\arguments{
\item{data}{set of texts to be written (obtained from
\code{\link{get.texts}})}

\item{groupBy1}{variable name (or vector of variable names) describing stylo
text class (texts grouping variable, e.g. author)}

\item{groupBy2}{variable name (or vector of variable names) describing stylo
text subclass (e.g. text title)}

\item{directory}{directory path to write corpus into}

\item{sample}{fraction of texts to sample from each group (from 0 to 1)}

\item{limit}{maximum number of characters in each file}
}
\value{
data.frame all texts which were not sampled
}
\description{
writes texts to a disk in a format supported by the stylo package
}
\examples{
\dontrun{
  # connect to the texts database
  db.connect()

  # fetch and filter texts
  allTexts = get.texts()
  interestingTexts = allTexts \%>\%
    filter(
      author \%in\% c('ABronte', 'CBronte', 'EBronte', 'Austen', 'Dickens',
        'Eliot', 'Fielding', 'Richardson', 'Sterne', 'Thackeray', 'Trollope')
    ) \%>\%
    collect()

  # write to disk in the stylo package format
  # (for the stylo() function, so in directory called "corpus")
  # taking the author as a main category and the title as a subcategory
  write.corpus(interestingTexts, 'author', 'title', 'corpus')

  # write to disk in the stylo package format
  # (for the classify() function, so there are two sets of text in
  #   directories "primary_set" and "secondary_set")
  # taking the author as a main category and the title as a subcategory
  primarySet = interestingTexts \%>\%
    group_by(author) \%>\%
    sample_frac(0.7) # assign 70\% of each uthors' texts to the primary set
  secondarySet = interestingTexts \%>\%
    anti_join(primarySet \%>\% select(text_id)) # all other goes to secondary set
  write.corpus(primarySet, 'author', 'title', 'primary_set')
  write.corpus(secondarySet, 'author', 'title', 'secondary_set')
}
}
\seealso{
\code{\link{get.texts}}
}

